PDF_Generator/Server_PM_ReportForm_PDF/requirements.txt
pyodbc==4.0.39
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
paho-mqtt==1.6.1
reportlab==4.0.4
Pillow==10.0.0
aiohttp>=3.8.0
asyncio